feat(AddDocument): refactor :)
- Se crea el FileDocumentFactory.
- Los archivos se descargan con el nombre original del archivo.
