+++ ClientApp/src/app/components/registry-details/registry-details.component.html
@@ -38,7 +38,7 @@
           </button>
         </td>
         <ng-template #file_template>
-            <button type="button" (click)="download(document.link)" class="btn btn-success">
+            <button type="button" (click)="download(document)" class="btn btn-success">
               <i class="glyphicon glyphicon-download-alt"></i>
             </button>
         </ng-template>
